OSDEV-2948 Extract facilities visibility token
